Move assertions with user related tests
