+++ apps/issuer/tests/test_assertion.py
@@ -11,15 +11,13 @@ import pytz
 import re
 from urllib import quote_plus
 
-from django.apps import apps
 from django.core import mail
 from django.core.urlresolvers import reverse
 from django.utils import timezone
 from oauth2_provider.models import Application
 
-from badgeuser.models import CachedEmailAddress
+from badgeuser.models import CachedEmailAddress, UserRecipientIdentifier
 from issuer.models import BadgeInstance, IssuerStaff, Issuer
-from mainsite.models import ApplicationInfo
 from mainsite.tests import BadgrTestCase, SetupIssuerHelper, SetupOAuth2ApplicationHelper
 from mainsite.utils import OriginSetting
 from rest_framework import serializers
@@ -1065,6 +1063,65 @@ class AssertionsChangedApplicationTests(SetupOAuth2ApplicationHelper, SetupIssue
         response = self.client.get('/v2/assertions/changed?since={}'.format(quote_plus(timestamp)))
         self.assertEqual(response.status_code, 200)
         self.assertEqual(len(response.data['result']), 0)
+
+
+class AssertionWithUserTests(SetupIssuerHelper, BadgrTestCase):
+    def test_assertion_has_user_post_issue(self):
+        # email and url identifiers
+        staff = self.setup_user(email='staff@example.com')
+        recipient = self.setup_user(email='recipient@example.com', authenticate=False)
+        issuer = self.setup_issuer(owner=staff)
+        badgeclass = self.setup_badgeclass(issuer=issuer)
+        award = badgeclass.issue(recipient_id=recipient.email)
+        self.assertEqual(award.user.pk, recipient.pk)
+
+    def test_assertion_user_with_verification(self):
+        staff = self.setup_user(email='staff@example.com')
+        issuer = self.setup_issuer(owner=staff)
+        badgeclass = self.setup_badgeclass(issuer=issuer)
+
+        recipient = self.setup_user(email='recipient@example.com', authenticate=False, verified=False)
+        award = badgeclass.issue(recipient_id=recipient.email)
+        self.assertEqual(award.user, None)
+
+        recipient2 = self.setup_user(email='recipient2example.com', authenticate=False)
+        award2 = badgeclass.issue(recipient_id=recipient2.email)
+        self.assertEqual(award2.user.pk, recipient2.pk)
+
+    def test_assertion_user_none_post_email_or_identifier_delete(self):
+        staff = self.setup_user(email='staff@example.com')
+        recipient = self.setup_user(email='recipient@example.com', authenticate=False)
+        issuer = self.setup_issuer(owner=staff)
+        badgeclass = self.setup_badgeclass(issuer=issuer)
+        badgeclass.issue(recipient_id=recipient.email)
+        award = BadgeInstance.objects.get(recipient_identifier=recipient.email)
+        self.assertEqual(award.user, recipient)
+
+        CachedEmailAddress.objects.get(email='recipient@example.com').delete()
+        award = BadgeInstance.objects.get(recipient_identifier=recipient.email)
+        self.assertEqual(award.user, None)
+
+        my_id = self.add_identifier(recipient, 'http://google.com')
+        badgeclass.issue(recipient_id='http://google.com', recipient_type=UserRecipientIdentifier.IDENTIFIER_TYPE_URL)
+        award = BadgeInstance.objects.get(recipient_identifier='http://google.com')
+        self.assertEqual(award.user, recipient)
+
+        my_id.delete()
+        award = BadgeInstance.objects.get(recipient_identifier=recipient.email)
+        self.assertEqual(award.user, None)
+
+    def test_user_cant_fetch_changed_assertions(self):
+        staff = self.setup_user(email='staff@example.com')
+        recipient = self.setup_user(email='recipient@example.com', authenticate=True)
+
+        issuer = self.setup_issuer(owner=staff)
+        badgeclass = self.setup_badgeclass(issuer=issuer)
+        badgeclass.issue(recipient_id=recipient.email)
+        badgeclass.issue(recipient_id=staff.email)
+        url = reverse('v2_api_assertions_changed_list')
+
+        response = self.client.get(url)
+        self.assertEqual(response.status_code, 404)
 
 
 class AllowDuplicatesAPITests(SetupIssuerHelper, BadgrTestCase):
